refactor: frontend components
