+++ src/components/layout/Nav.jsx
@@ -8,7 +8,7 @@ import FilterMenu from './dialogs/filters/FilterMenu'
 import UserOptions from './dialogs/UserOptions'
 import Tutorial from './dialogs/tutorial/Tutorial'
 import UserProfile from './dialogs/profile'
-import Search from './dialogs/Search'
+import Search from './dialogs/search'
 import MessageOfTheDay from './dialogs/Motd'
 import DonorPage from './dialogs/DonorPage'
 import Feedback from './dialogs/Feedback'
